moved RangeBar methods to a service
